+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- The 'filter' operator
+
 ### Changed
-Move `testutil` package out of `internal`
+- Move `testutil` package out of `internal`
 
 ## [0.9.10] - 2020-08-20
 ### Added